+++ NEWS.md
@@ -1,3 +1,27 @@
+# labelled 2.9.0
+
+**look_for() improvements:**
+
+* new function `look_for_and_select()` (#87)
+* `look_for()` can now search within factor levels and value labels (#104)
+
+**improvements for tagged NAs:**
+
+* better printing of value labels (#89)
+* new functions `user_na_to_tagged_na()`, `tagged_na_to_user_na()` and
+  `tagged_na_to_regular_na()`
+* new option `explicit_tagged_na` in `to_factor()` and `to_character()`
+* new functions `unique_tagged_na()`, `duplicated_tagged_na()`, 
+  `order_tagged_na()`, `sort_tagged_na()` (#90, #91)
+
+**other improvements:**
+
+* new functions `is_user_na()` and `is_regular_na()`
+* new set of unit tests (#99)
+* trying to apply a value label, `na_range()` or `na_values()` to a factor 
+  will now produce an error
+* bug fix in `foreign_to_labelled()` for Stata files (#100)
+
 # labelled 2.8.0
 
 * new helper `recode_if()` for recoding values based on condition, variable